requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro==1.9.0
distro-info===0.18ubuntu0.18.04.1
Django==1.11.11
dnspython==2.6.1
docutils==0.14
elevenlabs==1.4.1
email_validator==2.2.0
evdev==0.7.0
exceptiongroup==1.2.2
ez_setup==0.9
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
html5lib==0.999999999
httpcore==1.0.5
httplib2==0.9.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.9
markdown-it-py==3.0.0
Markups==2.0.1
MarkupSafe==2.1.5
mdurl==0.1.2
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
openai==1.35.14
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.8.2
pydantic_core==2.20.1
pyenchant==2.0.0
Pygments==2.18.0
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
python-apt==1.6.6
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.20
pytz==2018.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==6.0.1
reportlab==3.4.0
requests==2.32.3
requests-unixsocket==0.1.5
rich==13.7.1
roman==2.0.0
SecretStorage==2.3.1
shellingham==1.5.4
simplejson==3.13.2
six==1.11.0
sniffio==1.3.1
sqlparse==0.2.4
ssh-import-id==5.7
starlette==0.37.2
system-service==0.3
systemd-python==234
textile==3.0.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==15.1.0
wadllib==1.3.2
watchfiles==0.22.0
webencodings==0.5
websockets==12.0
xkit==0.0.0
zope.interface==4.3.2
